# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with authentication, roles and encryption.",
        "title": "Go API Dojo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users with optional filters and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "pending",
                            "active",
                            "suspended",
                            "banned"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name or email",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ListUsersResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get the authenticated user's profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.MeResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/login-history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get the login history of the authenticated user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of entries to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.LoginHistoryEntry"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Soft-delete a user account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required in cookie/dual mode",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "cannot delete another user's account",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Change the password of a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Password payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ChangePasswordRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Required in cookie/dual mode",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "current password is incorrect",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "cannot change another user's password",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/profile": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Profile payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateProfileRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Required in cookie/dual mode",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "cannot update another user's profile",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fault.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "a descriptive error message"
                }
            }
        },
...
        "handler.ListUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/handler.ListMeta"
                }
            }
        },
...
        "handler.LoginHistoryEntry": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.LoginStatus"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
...
        "handler.MeResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "description": "Profile",
                    "type": "string"
                },
                "headline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_2fa_enabled": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "language": {
                    "description": "Locale",
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "status": {
                    "$ref": "#/definitions/model.UserStatus"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vat_number": {
                    "description": "Sensitive — visible only to the owner",
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "address_line1": {
                    "description": "Address",
                    "type": "string"
                },
                "address_line2": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "banned_at": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "description": "Profile",
                    "type": "string"
                },
                "headline": {
                    "type": "string"
                },
                "id": {
                    "description": "Identity",
                    "type": "string"
                },
                "is_2fa_enabled": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "language": {
                    "description": "Locale",
                    "type": "string"
                },
                "last_login_at": {
                    "description": "Activity",
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "login_count": {
                    "type": "integer"
                },
                "longitude": {
                    "type": "number"
                },
                "postal_code": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "status": {
                    "$ref": "#/definitions/model.UserStatus"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}